+++ lib/screens/01_home/HomeScreen.dart
@@ -1,7 +1,8 @@
 import 'package:daangn/models/product.dart';
-import 'package:daangn/screens/01_home/components/product_detail.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+
+import 'components/product_item.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({
@@ -40,45 +41,15 @@ class HomeScreen extends StatelessWidget {
       body: ListView.separated(
         itemBuilder: (context, i) {
           /* 상품 카드 */
-          return PrductItem(
-            product: productList[i],
-          );
+          return PrductItem(product: productList[i]);
         },
-        separatorBuilder: (context, index) => Divider(),
+        separatorBuilder: (context, index) => Divider(
+          height: 0 /* 기본 높이가 있나 봄*/,
+          indent: 16,
+          endIndent: 16,
+          color: Colors.grey,
+        ),
         itemCount: productList.length,
-      ),
-    );
-  }
-}
-
-class PrductItem extends StatelessWidget {
-  const PrductItem({
-    super.key,
-    required this.product,
-  });
-
-  final Product product;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 135.0,
-      padding: EdgeInsets.all(16.0),
-      child: Row(
-        children: [
-          ClipRRect(
-            borderRadius: BorderRadius.circular(100.0),
-            child: Image.network(
-              product.urlToImage,
-              width: 115,
-              height: 115,
-              fit: BoxFit.cover,
-            ),
-          ),
-          SizedBox(width: 15.0),
-          /* 상품 티테일 */
-          ProductDetail(product: product)
-        ],
       ),
     );
   }
